restructure integration tests
