--- tests/integration/constant_sum.rs
@@ -1,6 +1,0 @@
-use crate::helpers::simulation_test;
-
-#[test]
-fn test_constant_sum() {
-    simulation_test("tests/circuits/constantSum.circom", [], [("0.out", 8)]);
-}
