--- tests/integration/helpers.rs
@@ -1,29 +1,0 @@
-use circom_2_arithc::{cli::Args, program::compile};
-use sim_circuit::{simulate, NumberU32};
-use std::collections::HashMap;
-
-pub fn simulation_test<
-    Input: IntoIterator<Item = (&'static str, u32)>,
-    Output: IntoIterator<Item = (&'static str, u32)>,
->(
-    test_file_path: &str,
-    input: Input,
-    expected_output: Output,
-) {
-    let compiler_input = Args::new(test_file_path.into(), "./".into());
-    let circuit = compile(&compiler_input).unwrap().build_circuit().unwrap();
-
-    let input = input
-        .into_iter()
-        .map(|(name, value)| (name.to_string(), NumberU32(value)))
-        .collect::<HashMap<String, NumberU32>>();
-
-    let expected_output = expected_output
-        .into_iter()
-        .map(|(name, value)| (name.to_string(), NumberU32(value)))
-        .collect::<HashMap<String, NumberU32>>();
-
-    let output = simulate(&circuit.to_sim(), &input).unwrap();
-
-    assert_eq!(output, expected_output);
-}
