--- tests/integration/infix_ops.rs
@@ -1,48 +1,0 @@
-use crate::helpers::simulation_test;
-
-#[test]
-fn test_infix_ops() {
-    simulation_test(
-        "tests/circuits/infixOps.circom",
-        [
-            ("0.x0", 0),
-            ("0.x1", 1),
-            ("0.x2", 2),
-            ("0.x3", 3),
-            ("0.x4", 4),
-            ("0.x5", 5),
-        ],
-        [
-            ("0.mul_2_3", 6),
-            // ("0.div_4_3", 1), // unsupported for NumberU32
-            ("0.idiv_4_3", 1),
-            ("0.add_3_4", 7),
-            ("0.sub_4_1", 3),
-            ("0.pow_2_4", 16),
-            ("0.mod_5_3", 2),
-            ("0.shl_5_1", 10),
-            ("0.shr_5_1", 2),
-            ("0.leq_2_3", 1),
-            ("0.leq_3_3", 1),
-            ("0.leq_4_3", 0),
-            ("0.geq_2_3", 0),
-            ("0.geq_3_3", 1),
-            ("0.geq_4_3", 1),
-            ("0.lt_2_3", 1),
-            ("0.lt_3_3", 0),
-            ("0.lt_4_3", 0),
-            ("0.gt_2_3", 0),
-            ("0.gt_3_3", 0),
-            ("0.gt_4_3", 1),
-            ("0.eq_2_3", 0),
-            ("0.eq_3_3", 1),
-            ("0.neq_2_3", 1),
-            ("0.neq_3_3", 0),
-            ("0.or_0_1", 1),
-            ("0.and_0_1", 0),
-            ("0.bit_or_1_3", 3),
-            ("0.bit_and_1_3", 1),
-            ("0.bit_xor_1_3", 2),
-        ],
-    );
-}
