--- tests/integration/main.rs
@@ -1,9 +1,0 @@
-mod add_zero;
-mod constant_sum;
-mod helpers;
-mod infix_ops;
-mod mat_elem_mul;
-mod prefix_ops;
-mod sum;
-mod under_constrained;
-mod x_eq_x;
