--- tests/integration/prefix_ops.rs
@@ -1,23 +1,0 @@
-use crate::helpers::simulation_test;
-
-#[test]
-#[should_panic] // FIXME: Should NOT panic (see comment below)
-fn test_prefix_ops() {
-    // FIXME: The compiler sees several of the outputs as inputs, leading to the error below
-    //        CircuitError(Inconsistency {
-    //            message: "Node 10 used for both input 0.complementC and output 0.complementC"
-    //        })
-    simulation_test(
-        "tests/circuits/prefixOps.circom",
-        [("0.a", 0), ("0.b", 1), ("0.c", 2)],
-        [
-            ("0.negateA", 0),                                          // -0
-            ("0.notA", 1),                                             // !0
-            ("0.notB", 0),                                             // !1
-            ("0.notC", 0),                                             // !2
-            ("0.complementA", 0b_11111111_11111111_11111111_11111111), // ~0
-            ("0.complementB", 0b_11111111_11111111_11111111_11111110), // ~1
-            ("0.complementC", 0b_11111111_11111111_11111111_11111101), // ~2
-        ],
-    );
-}
