--- tests/integration/sum.rs
@@ -1,10 +1,0 @@
-use crate::helpers::simulation_test;
-
-#[test]
-fn test_sum() {
-    simulation_test(
-        "tests/circuits/sum.circom",
-        [("0.a", 3), ("0.b", 5)],
-        [("0.out", 8)],
-    );
-}
